requirements.txt
fastapi[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
pydantic-settings
python-multipart